Figuring out the integrateIt function
I made a new repository and called it Metz-Midterm-2 for my second try.
I added n into the arguments to represent the number of partitions a
user would want for the Trapezoidal Rule function. I believe that it is
a necessary addition because it means the user can put any number of
partitions they want into the argument, which affects the value of the
h written in the function and thus the output itself. I will be adding
comments later and I am still unsure as to whether this function
properly works. It works for the specific equation I gave it but I will
ensure that it works for all functions soon.